Factor out writeToNav for use in 404
